extract git integration tests in separate file
